requirements.txt
Flask==3.1.2
python-dotenv==1.0.0
flask-cors==4.0.0

